+++ client/bindings/nodejs/types/bridge/client.ts
@@ -185,10 +185,6 @@ export interface __GetTipsIntervalMessage__ {
     name: 'getTipsInterval';
 }
 
-export interface __GetTokenSupplyMessage__ {
-    name: 'getTokenSupply';
-}
-
 export interface __GetProtocolParametersMessage__ {
     name: 'getProtocolParameters';
 }