+++ client/src/message_interface/message.rs
@@ -128,8 +128,6 @@ pub enum Message {
     GetMinPowScore,
     /// Returns the tips interval
     GetTipsInterval,
-    /// Returns the token supply
-    GetTokenSupply,
     /// Returns the protocol parameters
     GetProtocolParameters,
     /// Returns if local pow should be used or not